+++ CHANGELOG.md
@@ -11,3 +11,9 @@ Initial Release
 - In Track - Added line Options (line coor and line width)
 - In Track - Added last location dot
 - In Cluster and Heatmap - recenter the map in average location
+
+## v.1.2.1
+
+- In Track - remove pinch
+- In Track - add blink dot, include options (size $ color)
+- In Track - add option to remove auto center